requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.0.2
django-model-utils==4.0.0
djangorestframework==3.11.0
idna==2.9
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
